requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0

# Notion
notion-client==2.2.1

# Apify
apify-client==1.6.4

# Database
sqlalchemy==2.0.25
pandas==2.1.4

# Dashboard
streamlit==1.31.0
plotly==5.18.0

# Twitter/X API
tweepy==4.14.0

# LLM (for tweet generation)
anthropic==0.18.1
